feat(core): M2 슬롯·단일종류검증·정산 계산값 + 돈 long 전환
- 슬롯(칸) 계산 + 초과 시 출발 불가
- 견인 동물 단일종류 검증
- 정산 데이터 계산값 5개(이동시간·식량소모·출발적재·최종적정·과적비율)
- revenue/cost/netProfit long (+어댑터 임시 int 캐스팅)
- 출발 불가 사유 상세화, 콘솔 로그 무게/개수 구분
- 상태전환 중복·음수거리 방어, 죽은 코드 제거
- UI 값(IsOverloaded/LoadSpeedModifier/EstimatedFood) 노출

+++ Assets/_Project/11.CoreServices/Scripts/Save/SaveData.cs
@@ -237,7 +237,7 @@ namespace ND.Framework
         public string itemId = string.Empty;
         public string itemName = string.Empty;
         public float weight;
-        public int basePrice;
+        public long basePrice;
     }
 
     /// <summary>
